Sizepo

+++ sw/applications/0_DOOM/doom/doomstat.h
@@ -90,8 +90,8 @@ extern  boolean     autostart;
 
 // Selected by user. 
 extern  skill_t         gameskill;
-extern  int     gameepisode;
-extern  int     gamemap;
+extern  uint8_t         gameepisode;
+extern  uint8_t         gamemap;
 
 // If non-zero, exit the level after this number of minutes
 extern  int             timelimit;
